refactor(cauldron): deconstruct CauldronSubcomponents and move to molecules/cauldron
